requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-cloud-storage==2.19.0
python-dotenv==1.0.1
anthropic>=0.40.0
httpx>=0.27.0
pydantic>=2.0
